+++ app/scripts/data-fetchers/get-data-fetcher.js
@@ -2,7 +2,14 @@ import GBKDataFetcher from './genbank-fetcher';
 import LocalDataFetcher from './local-tile-fetcher';
 import DataFetcher from '../DataFetcher';
 
-const getDataFetcher = (dataConfig, pubSub) => {
+const getDataFetcher = (dataConfig, pubSub, pluginDataFetchers) => {
+  // Check if a plugin data fetcher is available.
+  const pluginDataFetcher = pluginDataFetchers[dataConfig.type];
+  if (pluginDataFetcher) {
+    // eslint-disable-next-line new-cap
+    return new pluginDataFetcher(dataConfig, pubSub);
+  }
+
   if (dataConfig.type === 'genbank') {
     return new GBKDataFetcher(dataConfig, pubSub);
   }